# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API pour le projet PEC2 Backend",
        "title": "API PEC2 Backend",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8090",
    "basePath": "/",
    "paths": {
...
        "/content-creators": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing content creator application (rejected or approved)",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content-creators"
                ],
                "summary": "Update a content creator application",
                "parameters": [
                    {
                        "type": "string",
                        "default": "My Creative Company",
                        "description": "Company name",
                        "name": "companyName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "SARL",
                        "description": "Company type",
                        "name": "companyType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "12345678901234",
                        "description": "SIRET number",
                        "name": "siretNumber",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "FR12345678901",
                        "description": "VAT number",
                        "name": "vatNumber",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "123 Business Street",
                        "description": "Street address",
                        "name": "streetAddress",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "75001",
                        "description": "Postal code",
                        "name": "postalCode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Paris",
                        "description": "City",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "France",
                        "description": "Country",
                        "name": "country",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "FR7630006000011234567890189",
                        "description": "IBAN",
                        "name": "iban",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "BNPAFRPP",
                        "description": "BIC",
                        "name": "bic",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Document proof (PDF, image)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "message: Application updated successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "403": {
                        "description": "error: Application cannot be updated",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "error: No application found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Submit an application to become a content creator",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content-creators"
                ],
                "summary": "Apply to become a content creator",
                "parameters": [
                    {
                        "type": "string",
                        "default": "My Creative Company",
                        "description": "Company name",
                        "name": "companyName",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "SARL",
                        "description": "Company type",
                        "name": "companyType",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "12345678901234",
                        "description": "SIRET number",
                        "name": "siretNumber",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "FR12345678901",
                        "description": "VAT number",
                        "name": "vatNumber",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "default": "123 Business Street",
                        "description": "Street address",
                        "name": "streetAddress",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "75001",
                        "description": "Postal code",
                        "name": "postalCode",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "Paris",
                        "description": "City",
                        "name": "city",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "France",
                        "description": "Country",
                        "name": "country",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "FR7630006000011234567890189",
                        "description": "IBAN",
                        "name": "iban",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "default": "BNPAFRPP",
                        "description": "BIC",
                        "name": "bic",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "Document proof (PDF, image)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "message: Application submitted successfully",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "error: Invalid input",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "409": {
                        "description": "error: Application already exists",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "error: Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/content-creators/all": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all content creator applications (Admin access only)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "content-creators"
                ],
                "summary": "Get all content creator applications (Admin)",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.ContentCreatorInfo"
                            }
                        }
                    },
                    "401": {
                        "description": "error: Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "403": {
                        "description": "error: Forbidden - Admin access required",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "error: Error message",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.ContentCreatorInfo": {
            "type": "object",
            "required": [
                "bic",
                "city",
                "companyName",
                "companyType",
                "country",
                "documentProofUrl",
                "iban",
                "postalCode",
                "siretNumber",
                "streetAddress"
            ],
            "properties": {
                "bic": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "companyName": {
                    "type": "string"
                },
                "companyType": {
                    "type": "string"
                },
                "country": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "documentProofUrl": {
                    "type": "string"
                },
                "iban": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "postalCode": {
                    "type": "string"
                },
                "siretNumber": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/models.ContentCreatorStatusType"
                },
                "streetAddress": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "vatNumber": {
                    "type": "string"
                }
            }
        },
        "models.ContentCreatorStatusType": {
            "type": "string",
            "enum": [
                "PENDING",
                "APPROVED",
                "REJECTED"
            ],
            "x-enum-varnames": [
                "ContentCreatorStatusPending",
                "ContentCreatorStatusApproved",
                "ContentCreatorStatusRejected"
            ]
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Entrez le JWT avec le préfixe Bearer: Bearer \u003cJWT\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}